+++ simas-be/database/migrations/2022_12_13_023439_add_kromium_field.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('kualitas_airs', function (Blueprint $table) {
-           $table->float('kromium6')->nullable();
+           $table->double('kromium6')->nullable();
         });
     }
